extracted comment section into own component
